feat: prepare in verify if env var is set
It's common to build assets that need updated package metadata, such as
a CLI with a `--version` flag. If the package metadata is not updated
before building, this doesn't work.

To support this use-case we look for a `PREPARE_IN_VERIFY` environment
variable and, if set, do the normal prepare steps but as part of verify.

Since verify is still run on `dryRun`, that is how users can accomplish
updating `package.yaml` before building the assets (without doing any
other release activities yet), then they can release for real after.

Additionally, we make the real `prepare` function idempotent in case
we're being used this way, though it would not be particularly
problematic to re-update the `package.yaml` unnecessarily.